tests: Add an Undefined test suite
