+++ tests/unit/descriptors.cpp
@@ -1865,52 +1865,6 @@ TEST_F(NegativeDescriptors, DSBufferInfoTemplate) {
     m_errorMonitor->VerifyFound();
 }
 
-TEST_F(NegativeDescriptors, BindInvalidPipelineLayout) {
-    TEST_DESCRIPTION("https://github.com/KhronosGroup/Vulkan-ValidationLayers/issues/6621");
-    RETURN_IF_SKIP(Init());
-    InitRenderTarget();
-
-    vkt::Buffer buffer(*m_device, 32, VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT);
-    OneOffDescriptorSet descriptor_set(m_device,
-                                       {{0, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, 1, VK_SHADER_STAGE_FRAGMENT_BIT, nullptr}});
-    descriptor_set.WriteDescriptorBufferInfo(0, buffer, 0, VK_WHOLE_SIZE);
-    descriptor_set.UpdateDescriptorSets();
-
-    // Create PSO to be used for draw-time errors below
-    const char *fsSource = R"glsl(
-        #version 450
-        layout(location=0) out vec4 x;
-        layout(set=0, binding=1) uniform foo1 { vec4 y; };
-        void main(){
-           x = y;
-        }
-    )glsl";
-    VkShaderObj vs(this, kVertexMinimalGlsl, VK_SHADER_STAGE_VERTEX_BIT);
-    VkShaderObj fs(this, fsSource, VK_SHADER_STAGE_FRAGMENT_BIT);
-
-    CreatePipelineHelper pipe(*this);
-    pipe.shader_stages_ = {vs.GetStageCreateInfo(), fs.GetStageCreateInfo()};
-    pipe.pipeline_layout_ = vkt::PipelineLayout(*m_device, {&descriptor_set.layout_});
-    // consume VU so we create a pipeline handle
-    m_errorMonitor->SetAllowedFailureMsg("VUID-VkGraphicsPipelineCreateInfo-layout-07988");
-    pipe.CreateGraphicsPipeline();
-    if (pipe == VK_NULL_HANDLE) {
-        GTEST_SKIP() << "Driver failed to create a invalid pipeline handle";
-    }
-
-    m_command_buffer.Begin();
-    m_command_buffer.BeginRenderPass(m_renderPassBeginInfo);
-    vk::CmdBindPipeline(m_command_buffer, VK_PIPELINE_BIND_POINT_GRAPHICS, pipe);
-    vk::CmdBindDescriptorSets(m_command_buffer, VK_PIPELINE_BIND_POINT_GRAPHICS, pipe.pipeline_layout_, 0, 1, &descriptor_set.set_,
-                              0, nullptr);
-    m_errorMonitor->SetDesiredError("VUID-vkCmdDraw-None-08114");
-    vk::CmdDraw(m_command_buffer, 1, 0, 0, 0);
-    m_errorMonitor->VerifyFound();
-
-    m_command_buffer.EndRenderPass();
-    m_command_buffer.End();
-}
-
 TEST_F(NegativeDescriptors, ConstantArrayElementNotBound) {
     AddRequiredFeature(vkt::Feature::vertexPipelineStoresAndAtomics);
     RETURN_IF_SKIP(Init());
